Death animation is functional. Still missing transition. Cleaned up code for death.
Still missing transition screen because I haven't figured it out. Cleaned up all the garbage I created with the code (all the if / else statements) and turned the death animation into a proper finite state machine instead of being ...that. Means we can add more steps to the death as well and also the timers are a bit easier to adjust / know which is which since they come in order.